update fields using positivesmallint

+++ api/ellandi/registration/models.py
@@ -328,7 +328,7 @@ class Learning(TimeStampedModel):
     name = models.CharField(max_length=255, blank=True, null=True)
     duration_minutes = models.PositiveIntegerField(null=True)
     date_completed = models.DateField(null=True)
-    cost_pounds = models.PositiveSmallIntegerField(blank=True, null=True)
+    cost_pounds = models.PositiveIntegerField(blank=True, null=True)
     cost_unknown = models.BooleanField(blank=True, null=True)
 
 
@@ -394,7 +394,7 @@ class Course(models.Model):
     short_description = models.CharField(max_length=1024, blank=True, null=True)
     long_description = models.TextField(blank=True, null=True)
     status = models.CharField(max_length=128, choices=Status.choices, blank=True, null=True)
-    cost_pounds = models.PositiveSmallIntegerField(blank=True, null=True)
+    cost_pounds = models.PositiveIntegerField(blank=True, null=True)
     duration_minutes = models.PositiveIntegerField(blank=True, null=True)
     private = models.BooleanField(blank=True, null=True)
     course_type = models.CharField(max_length=256, choices=CourseType.choices, blank=True, null=True)
